refactor: 4: standardise site-specific override convention (#1416)
move ga2 view + content overrides into `/ga2/` subdirectories and drop
the `*ga2` file suffixes, matching the dominant override convention:

- roadmapviewga2.tsx -> roadmapview/ga2/roadmapview.tsx
- partnerresourcesviewga2.tsx -> partnerresourcesview/ga2/partnerresourcesview.tsx
- section{about,roadmap}ga2.mdx -> content/ga2/section{about,roadmap}.mdx
- sectionsubheroga2.mdx -> home/content/ga2/sectionsubhero.mdx

exported symbols drop the ga2 suffix (directory is the disambiguator);
new ga2/index.tsx sub-barrels re-export the content, and importers are
updated directly (no shims), so no `*ga2` suffixes remain.
