+++ app/components/Home/components/Section/components/ga2/SectionSubHero/sectionSubHero.tsx
@@ -4,7 +4,7 @@ import {
   SLIDE_PROPS,
 } from "@/components/Home/components/Section/components/SectionSubHero/constants";
 import { useAutoCycle } from "@/components/Home/components/Section/components/SectionSubHero/hooks";
-import { SubHeroContentGA2 } from "@/components/Home/content";
+import { SubHeroContent } from "@/components/Home/content/ga2";
 import { AccordionDetails, AccordionSummary, Fade, Slide } from "@mui/material";
 import { JSX } from "react";
 import { ACCORDION, BUTTON, IMAGE } from "./instructions";
@@ -30,7 +30,7 @@ export const SectionSubHero = (): JSX.Element => {
       <SectionLayout>
         <SectionSubLayout>
           <Subhead>
-            <SubHeroContentGA2 />
+            <SubHeroContent />
           </Subhead>
           <AccordionBox>
             {Object.entries(ACCORDION).map(([value, { details, title }]) => (
