+++ app/components/Home/content/ga2/index.tsx
