+++ app/components/content/ga2/index.tsx
@@ -1,3 +1,2 @@
 export { default as SectionAbout } from "./sectionAbout.mdx";
 export { default as SectionRoadmap } from "./sectionRoadmap.mdx";
-export { default as SectionVision } from "./sectionVision.mdx";
