+++ app/components/content/ga2/sectionAbout.mdx
