+++ pages/about/partner-resources/index.tsx
@@ -1,8 +1,8 @@
 import { getPageMeta } from "@/common/meta/utils";
 import { StyledPagesMain } from "@/components/Layout/components/Main/main.styles";
 import { config } from "@/config/config";
+import { PartnerResourcesView as PartnerResourcesViewGA2 } from "@/views/PartnerResourcesView/ga2/partnerResourcesView";
 import { PartnerResourcesView } from "@/views/PartnerResourcesView/partnerResourcesView";
-import { PartnerResourcesViewGA2 } from "@/views/PartnerResourcesView/partnerResourcesViewGA2";
 import { APP_KEYS } from "@site-config/common/constants";
 import { GetStaticProps } from "next";
 import { JSX } from "react";
